Refactor toolbar file structure
